+++ index.js
@@ -166,6 +166,8 @@ export { default as BalloonComponent } from './src/featureInfo/BalloonComponent.
 export { default as AddressBalloonComponent } from './src/featureInfo/AddressBalloonComponent.vue';
 export { default as IframeFeatureInfoView } from './src/featureInfo/iframeFeatureInfoView.js';
 export { default as TableFeatureInfoView } from './src/featureInfo/tableFeatureInfoView.js';
+export { default as MarkdownFeatureInfoView } from './src/featureInfo/markdownFeatureInfoView.js';
+export { default as MarkdownBalloonFeatureInfoView } from './src/featureInfo/markdownBalloonFeatureInfoView.js';
 export {
   getHighlightStyle,
   featureInfoViewSymbol,