+++ src/featureInfo/markdownFeatureInfoView.js
@@ -1,5 +1,8 @@
 import AbstractFeatureInfoView from './abstractFeatureInfoView.js';
-import { parseAndSanitizeMarkdown } from '../application/markdownHelper.js';
+import {
+  parseAndSanitizeMarkdown,
+  replaceAttributes,
+} from '../application/markdownHelper.js';
 
 /**
  * @description A component to render markdown html
@@ -22,64 +25,8 @@ const MarkdownComponent = {
  */
 
 /**
- * @typedef {import("./abstractFeatureInfoView.js.js").FeatureInfoProps & { html: string }} MarkdownFeatureInfoViewProps
+ * @typedef {import("./abstractFeatureInfoView.js").FeatureInfoProps & { html: string }} MarkdownFeatureInfoViewProps
  */
-
-/**
- * @param {Record<string, unknown>} parent
- * @param {(string|number)[]} keys
- * @returns {undefined|T}
- * @template {*} T
- */
-function findRecursive(parent, keys) {
-  if (keys.length === 1) {
-    return parent[keys[0]];
-  } else {
-    const nextKey = keys.shift();
-    const nextParent = parent[nextKey];
-    if (nextParent) {
-      return findRecursive(nextParent, keys);
-    }
-  }
-  return undefined;
-}
-
-/**
- * @param {string|string[]} template
- * @param {Record<string, unknown>} attributes
- * @returns {string}
- */
-function replaceAttributes(template, attributes) {
-  const templateString = Array.isArray(template)
-    ? template.join('\n')
-    : template;
-  return templateString.replace(/\{\{([^}]+)}}/g, (p, value) => {
-    const keys = value.trim().split('.');
-
-    for (let i = 0; i < keys.length; i++) {
-      let key = keys[i];
-      if (typeof key === 'string') {
-        const indices = [];
-        let arrayIndex = /\[["']?([^\]]+)["']?]$/.exec(key);
-        while (arrayIndex != null) {
-          let bracketKey = arrayIndex[1];
-          if (/^\d+$/.test(bracketKey)) {
-            bracketKey = Number(bracketKey);
-          }
-          indices.push(bracketKey);
-          key = key.substring(0, arrayIndex.index);
-          arrayIndex = /\[["']?([^\]]+)["']?]$/.exec(key);
-        }
-
-        if (indices.length > 0) {
-          keys.splice(i, 1, key, ...indices);
-        }
-      }
-    }
-
-    return findRecursive(attributes, keys) ?? '';
-  });
-}
 
 /**
  * @class